+++ GridKit/Model/PhasorDynamics/Converter/REECB/Reecb.cpp
@@ -1,0 +1,27 @@
+/**
+ * @file Reecb.cpp
+ * @author Luke Lowery (lukel@tamu.edu)
+ * @brief Non-Enzyme instantiation for the REECB electrical-control model.
+ */
+
+#include "ReecbImpl.hpp"
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    namespace Converter
+    {
+      template <class ScalarT, typename IdxT>
+      int Reecb<ScalarT, IdxT>::evaluateJacobian()
+      {
+        Log::misc() << "Evaluate Jacobian for Reecb..." << std::endl;
+        Log::misc() << "Jacobian evaluation is not implemented!" << std::endl;
+        return 0;
+      }
+
+      template class Reecb<double, long int>;
+      template class Reecb<double, size_t>;
+    } // namespace Converter
+  } // namespace PhasorDynamics
+} // namespace GridKit